refactor(radio): consolidate formatTimeWithSeconds in dateTime util with tests
