+++ assets/src/components/radio/rtts/types.ts
@@ -26,19 +26,7 @@ export interface RttCall {
   status: RttStatus
 }
 
-export const formatTimeWithSeconds = (
-  dateOrStr?: Date | string | null
-): string => {
-  if (!dateOrStr) return ""
-  const d = typeof dateOrStr === "string" ? new Date(dateOrStr) : dateOrStr
-  if (isNaN(d.getTime())) return ""
-  const hours24 = d.getHours()
-  const hours12 = hours24 % 12 || 12
-  const mins = d.getMinutes().toString().padStart(2, "0")
-  const secs = d.getSeconds().toString().padStart(2, "0")
-  const ampm = hours24 >= 12 ? "PM" : "AM"
-  return `${hours12}:${mins}:${secs} ${ampm}`
-}
+export { formatTimeWithSeconds } from "../../../util/dateTime"
 
 export const CALL_TYPE_PRIORITY: Record<RttCallType, number> = {
   Emergency: 1,
